Revert "Extract getTabs for EntityPickerModal"
This reverts commit 4e875a1fc125b7bd11fc869c8d2ee48acae39983.
